refactor: split core logic from CLI to support a future API frontend
Extracts framework-agnostic orchestration (models, service, errors) into
zenflow.core so a FastAPI layer can reuse init_project() without depending
on argparse/input()/sys.exit. zenflow.cli becomes a thin wizard adapter
over that service.
