+++ tests/test_deployment.py
@@ -7,7 +7,7 @@ from pathlib import Path
 
 import pytest
 
-from zenflow.deployment import (
+from zenflow.core.deployment import (
     deploy_agents,
     deploy_guidelines_to_github,
     deploy_guidelines_to_skills,
